refactor(github): share api auth error mapping
