+++ src/main/core/github/services/issue-service.ts
@@ -2,7 +2,7 @@ import type { Octokit } from '@octokit/rest';
 import type { IssueListError } from '@shared/issue-providers';
 import { isGitHubDotComHost, type RepositoryRef } from '@shared/repository-ref';
 import { err, ok, type Result } from '@shared/result';
-import type { GitHubApiAuthError } from './github-api-auth-service';
+import type { GitHubApiAuthError } from './github-api-auth-errors';
 import { getOctokit } from './octokit-provider';
 
 // ---------------------------------------------------------------------------
